grpc-test-vkr-main/vkr-glossary-grpc-project/glossary-grpc/web-service/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
grpcio ~= 1.30
grpcio-tools ~= 1.30

